requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
numpy>=1.26.0
pillow>=10.4.0
opencv-python>=4.10.0
pytest>=8.2.0
httpx>=0.27.0
